VBoxGlobal.cpp: size_t warning (Qt seems to use int, not size_t).
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@18495 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/VBoxGlobal.cpp
@@ -1653,7 +1653,7 @@ QString VBoxGlobal::detailsReport (const CMachine &aMachine, bool aIsNewVM,
         int rows = 2; /* including section header and footer */
 
         CHardDiskAttachmentVector vec = aMachine.GetHardDiskAttachments();
-        for (size_t i = 0; i < (size_t) vec.size(); ++ i)
+        for (int i = 0; i < vec.size(); ++ i)
         {
             CHardDiskAttachment hda = vec [i];
             CHardDisk hd = hda.GetHardDisk();
